Fixed mouse click detection + other bugfixes and refactoring

+++ ScreenRecorderLib/mouse_pointer.h
@@ -42,9 +42,9 @@ public:
 
 	HRESULT Initialize(ID3D11DeviceContext *ImmediateContext, ID3D11Device *Device);
 
-	HRESULT DrawMousePointer(_In_ PTR_INFO* PtrInfo, _In_ ID3D11DeviceContext* DeviceContext _In_, ID3D11Device* Device, _In_ ID3D11Texture2D* bgTexture, DXGI_MODE_ROTATION rotation);
-	HRESULT DrawMouseClick(_In_ PTR_INFO* PtrInfo, _In_ ID3D11Texture2D* bgTexture, std::string colorStr, int radius, DXGI_MODE_ROTATION rotation);
-	HRESULT GetMouse(_Inout_ PTR_INFO* PtrInfo, _In_ DXGI_OUTDUPL_FRAME_INFO* FrameInfo, RECT screenRect, IDXGIOutputDuplication* DeskDupl);
+	HRESULT DrawMousePointer(_In_ PTR_INFO *PtrInfo, _In_ ID3D11DeviceContext *DeviceContext, _In_ ID3D11Device *Device, _In_ ID3D11Texture2D *bgTexture, DXGI_MODE_ROTATION rotation);
+	HRESULT DrawMouseClick(_In_ PTR_INFO *PtrInfo, _In_ ID3D11Texture2D *bgTexture, std::string colorStr, float radius, DXGI_MODE_ROTATION rotation);
+	HRESULT GetMouse(_Inout_ PTR_INFO *PtrInfo, _In_ DXGI_OUTDUPL_FRAME_INFO *FrameInfo, RECT screenRect, IDXGIOutputDuplication* DeskDupl);
 	void CleanupResources();
 private:
 #pragma region Mouse Drawing
@@ -64,9 +64,9 @@ private:
 
 	long ParseColorString(std::string color);
 	float GetCurrentDpi();
-	void GetPointerPosition(_In_ PTR_INFO* PtrInfo, DXGI_MODE_ROTATION rotation,int desktopWidth, int desktopHeight, _Out_ INT* PtrLeft, _Out_ INT* PtrTop);
-	HRESULT ProcessMonoMask(_In_ ID3D11Texture2D* bgTexture, _In_ ID3D11DeviceContext* DeviceContext, _In_ ID3D11Device* Device, DXGI_MODE_ROTATION rotation, bool IsMono, _Inout_ PTR_INFO* PtrInfo, _Out_ INT* PtrWidth, _Out_ INT* PtrHeight, _Out_ INT* PtrLeft, _Out_ INT* PtrTop, _Outptr_result_bytebuffer_(*PtrHeight * *PtrWidth * BPP) BYTE** InitBuffer, _Out_ D3D11_BOX* Box);
-	HRESULT InitShaders(ID3D11DeviceContext* DeviceContext, ID3D11Device* Device);
+	void GetPointerPosition(_In_ PTR_INFO *PtrInfo, DXGI_MODE_ROTATION rotation, int desktopWidth, int desktopHeight, _Out_ INT *PtrLeft, _Out_ INT *PtrTop);
+	HRESULT ProcessMonoMask(_In_ ID3D11Texture2D *bgTexture, _In_ ID3D11DeviceContext *DeviceContext, _In_ ID3D11Device *Device, DXGI_MODE_ROTATION rotation, bool IsMono, _Inout_ PTR_INFO *PtrInfo, _Out_ INT *PtrWidth, _Out_ INT *PtrHeight, _Out_ INT *PtrLeft, _Out_ INT *PtrTop, _Outptr_result_bytebuffer_(*PtrHeight * *PtrWidth * BPP) BYTE **InitBuffer, _Out_ D3D11_BOX *Box);
+	HRESULT InitShaders(ID3D11DeviceContext *DeviceContext, ID3D11Device *Device);
 	HRESULT InitMouseClickTexture(ID3D11DeviceContext *ImmediateContext, ID3D11Device *Device);
 };
